Add softReset function.
Experiments with using the WDT to reset the MCU failed.  This might
have something to do with the interaction with the LowPower library,
but I'm not sure.  After reading the datasheet and looking over the
code, I was unable to determine why the MCU would go into an infinite
reboot loop

I might investigate this further in the future, but this function seems
to work for now.