Update configure packet(s).
Refactors the configure packets to a single configure packet that
contains a series of offset/value pairs.  This makes it easier to
change individual bytes in the configuration (and it's shorter to
transmit when only changing a few bytes).

This also updates saveConfiguration to only write bytes that differ,
whereas before it would write entire structs if only a single byte
differed.